Clear button added.
This is a slightly longer description of what I added with this commit. The assignment calls for a clear button, built using Bootstrap that when clicked, clear the DOM within the messages div container. If that container is empty, the button is disabled. There's a default attribute for buttons called disabled in Bootstrap so this was fairly straight-forward.